added pagination

+++ apps/api/src/post/post.resolver.ts
@@ -1,4 +1,4 @@
-import { Resolver, Query, Mutation, Args, Int, Context } from '@nestjs/graphql';
+import { Resolver, Query, Mutation, Args, Int, Context, Float } from '@nestjs/graphql';
 import { PostService } from './post.service';
 import { Post } from './entities/post.entity';
 import { CreatePostInput } from './dto/create-post.input';
@@ -19,8 +19,8 @@ export class PostResolver {
   @Query(() => [Post], { name: 'posts' })
   findAll(
     @Context() context,
-    @Args('skip', { type: () => Int, nullable: true }) skip?: number,
-    @Args('take', { type: () => Int, nullable: true }) take?: number,
+    @Args('skip', { type: () => Float, nullable: true }) skip?: number,
+    @Args('take', { type: () => Float, nullable: true }) take?: number,
   ) {
     const user = context.req.user; // Access the authenticated user from the request
     return this.postService.findAll({skip,take});
